Remove _flask_impl workaround.
Pylint seems to grok it now.
